+++ resources/views/layouts/navigation-simple.blade.php
@@ -23,7 +23,11 @@
 
                     <x-nav.link url="{{ route('customer.search') }}" :name="__('Kundensuche')"><x-svg.search class="h-6 w-6" /> </x-nav.link>
                     <x-nav.link url="{{ route('search.global') }}" :name="__('Globale Suche')"><x-svg.db class="h-6 w-6" /> </x-nav.link>
-                    <x-nav.link url="{{ route('search.remote') }}" :name="__('Rustdesk Suche')" target="_blank"><x-svg.software.rustdesk class="h-6 w-6" /> </x-nav.link>
+                    {{-- Nur mit dem Recht: Die Route verlangt es, der Verweis
+                         fuehrte sonst auf eine 403-Seite. --}}
+                    @can('remote_search')
+                        <x-nav.link url="{{ route('search.remote') }}" :name="__('Rustdesk Suche')" target="_blank"><x-svg.software.rustdesk class="h-6 w-6" /> </x-nav.link>
+                    @endcan
 
                 </div>
             @endcannot